WIP: convert "text/coffeescript" to "module" at runtime
Coffeescript supports modules, but its in-browser compiler uses
<script type="text/coffeescript">, and ESM imports need
<script type="module"> to work in the browser.

I thought I would try rolling my own solution, compiling coffeescript
and running it as a <script type="module">.

The importmap is able to point it at the correct files, but, of course,
the coffeescript depends on other coffeescript files, which the browser
doesn't know how to handle.

(Also the skele2d.js is UMD, not ESM, so it wouldn't work once it gets
to that stage, although I could make another build for ESM.)

I think I should either convert the vanilla example to plain JS, or
drop the usage of import/export, and just use globals.
(I could also imagine creating a service worker to intercept requests
for .coffee files and compile them, but that seems way out of scope.)