Update readme, to reflect the different examples
Also, having multiple examples makes me confident enough to weaken the
warning that you probably shouldn't try to use this library yet.
Maybe it's because I'm working on it right now, maybe it's just a mood.
I just did a hell of a lot of maintenance on it yesterday, so, I mean,
it FEELS MAINTAINED, at the moment, from my end, from my perspective.