Reorganize readme
The Dev Setup section was a mess, so I broke it up into sub-sections.
I may have fixed some problems with commands, and where to run them.
I also added a bit more blurb at the top of the readme.

I think it got messy partly because writing about the development setup
but I _want_ to be writing documentation for users of the library.
So it got a little muddled regarding the audience.