Maybe something like this could work?
I'm going in reverse order, which means that the first splice I do will be at the end,
which with cyclical splicing will immediately change the indices, which was what I was avoiding by going in reverse order.
I'll need to delay the cyclical part of the splicing until the end,
or handle an offset for splicing.