Find strands in a new way, to keep certain segments separate
It previously was joining strands of any points that are contiguous
in their indices (cyclically). This makes sense for points that are
within the brush radius, but if a point is outside of the brush radius,
with segments coming off of it that intersect the brush, it should not
join these, otherwise the point outside will be lost.

With this new approach, the point will be kept as a little triangle
jutting out from the brush.
Finally, all of these cases work*, and work regardless of which points
are first/last, in testing on a triangle:
- Brush encompassing one point
- Brush intersecting a line segment with no points encompassed
- Brush intersecting two line segments with no points encompassed

*Except for deciding which arc to use, which is still buggy.