Move tools bar to the bottom
This is much more compact and out of the way.
Before, when double clicking to edit an entity, it would often appear
over top of the entity if the entity was towards the right, which is
how you would position the view because the Entities palette is on the
left. So it would be obscured even though the Entities palette would
then hide, making room on the left that would have lead to you not
positioning the view like that, which made it extra annoying.

Now the layout is much more reasonable, although it wants for icons,
and the brush size slider label looks pretty cramped in there.

To make this layout change, I had to put `pointer-events: none` on
the editor, in order to use flexbox rather than absolute positioning.
This makes things awkward, as I have to re-enable interactivity on
each child that needs it. I've avoided this before due to the
maintenance burden, but I'm giving in, for the much-improved layout.
The problem with absolute positioning is that I would need to fix the
height of the tools bar as a constant in order to calculate an offset
in order to make the sidebars not be occluded by the tools bar.