Add PIXI.js example
It's very hacky how I'm cleaning up PIXI.js objects, but I'm bringing
this code into Skele2D as an example so I can work on the Skele2D API
improvements to support using an alternative renderer, more easily than
if I were to keep this code in an external project and work on it with
npm link skele2d. Or at least I can have a clearer git history, that is.